fix: moved faq to feature module
