+++ feature/faq/src/main/res/values/strings.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="feature_faq_question1">How can I add an Account?</string>
+    <string name="feature_faq_question2">How can I access my Payment History?</string>
+    <string name="feature_faq_question3">How to make a Transfer?</string>
+    <string name="feature_faq_question4">How can I add a Card?</string>
+    <string name="feature_faq_answer1">To add an account, navigate to finance section and click on the add account button. Then select your bank and verify your mobile number to get going. </string>
+    <string name="feature_faq_answer2">Navigate to Payments section. You will find your payment history under the History tab.</string>
+    <string name="feature_faq_answer3">To make a transfer, navigate to Payments section.Under the Send tab, from there you can choose the type of transfer, add the amount and submit.</string>
+    <string name="feature_faq_answer4">Navigate to Finance section. Click on Add Card under the Cards tab.</string>
+    <string name="feature_faq_frequently_asked_questions">Frequently Asked Question</string>
+</resources>
