+++ mifospay/src/main/res/values/strings.xml
@@ -122,14 +122,6 @@
     <string name="label_select_picture">Select picture</string>
 
 
-    <string name="question1">How can I add an Account?</string>
-    <string name="question2">How can I access my Payment History?</string>
-    <string name="question3">How to make a Transfer?</string>
-    <string name="question4">How can I add a Card?</string>
-    <string name="answer1">To add an account, navigate to finance section and click on the add account button. Then select your bank and verify your mobile number to get going. </string>
-    <string name="answer2">Navigate to Payments section. You will find your payment history under the History tab.</string>
-    <string name="answer3">To make a transfer, navigate to Payments section.Under the Send tab, from there you can choose the type of transfer, add the amount and submit.</string>
-    <string name="answer4">Navigate to Finance section. Click on Add Card under the Cards tab.</string>
     <string name="setup_upi">Setup UPI</string>
     <string name="change_upi_pin">Change UPI PIN</string>
     <string name="forgot_upi_pin">Forgot UPI PIN</string>
