+++ settings.gradle.kts
@@ -37,3 +37,4 @@ include(":feature:passcode")
 include(":feature:auth")
 include(":feature:make-transfer")
 include(":feature:send-money")
+include(":feature:faq")
